+++ src/main/java/com/daagn/clonestudy/common/RepositoryForEmTest.java
@@ -1,8 +1,7 @@
-package com.daagn.clonestudy;
+package com.daagn.clonestudy.common;
 
+import com.daagn.clonestudy.member.domain.Member;
 import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
