+++ src/main/java/com/daagn/clonestudy/member/domain/MannerEvaluationRepository.java
@@ -1,5 +1,6 @@
-package com.daagn.clonestudy;
+package com.daagn.clonestudy.member.domain;
 
+import com.daagn.clonestudy.member.domain.MannerEvaluation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MannerEvaluationRepository extends JpaRepository<MannerEvaluation, Long> {
