+++ src/main/java/com/daagn/clonestudy/member/domain/Member.java
@@ -1,5 +1,6 @@
-package com.daagn.clonestudy;
+package com.daagn.clonestudy.member.domain;
 
+import com.daagn.clonestudy.common.BaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
