+++ src/main/java/com/daagn/clonestudy/member/domain/MemberRepository.java
@@ -1,5 +1,6 @@
-package com.daagn.clonestudy;
+package com.daagn.clonestudy.member.domain;
 
+import com.daagn.clonestudy.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
