+++ src/main/java/com/daagn/clonestudy/post/domain/PostImage.java
@@ -1,5 +1,6 @@
-package com.daagn.clonestudy;
+package com.daagn.clonestudy.post.domain;
 
+import com.daagn.clonestudy.post.domain.Post;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
