+++ src/main/java/com/daagn/clonestudy/post/domain/PostImageRepository.java
@@ -1,0 +1,8 @@
+package com.daagn.clonestudy.post.domain;
+
+import com.daagn.clonestudy.post.domain.PostImage;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostImageRepository extends JpaRepository<PostImage, Long> {
+
+}
