+++ src/main/java/com/daagn/clonestudy/post/domain/Purchase.java
@@ -1,5 +1,7 @@
-package com.daagn.clonestudy;
+package com.daagn.clonestudy.post.domain;
 
+import com.daagn.clonestudy.member.domain.Member;
+import com.daagn.clonestudy.post.domain.Post;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -8,39 +10,33 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
 
 @Entity
 @NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class MannerEvaluation {
+public class Purchase {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
+  @Column(nullable = false)
+  private Boolean confirmed;
+
   @JoinColumn(nullable = false)
   @ManyToOne(fetch = FetchType.LAZY)
-  @OnDelete(action = OnDeleteAction.CASCADE)
-  private Member receiver;
+  private Member buyer;
 
   @JoinColumn(nullable = false)
   @ManyToOne(fetch = FetchType.LAZY)
-  @OnDelete(action = OnDeleteAction.CASCADE)
-  private Member sender;
+  private Post post;
 
-  @Column(nullable = false)
-  private String content;
-
-  @Column(nullable = false)
-  private Boolean isBuyer;
-
-  @Column(nullable = false)
-  private Boolean isPositive;
+  @Builder
+  public Purchase(Member buyer, Post post){
+    this.confirmed = false;
+    this. buyer = buyer;
+    this.post = post;
+  }
 
 }
