+++ src/main/java/com/daagn/clonestudy/post/domain/PurchaseRepository.java
@@ -1,5 +1,6 @@
-package com.daagn.clonestudy;
+package com.daagn.clonestudy.post.domain;
 
+import com.daagn.clonestudy.post.domain.Purchase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PurchaseRepository extends JpaRepository<Purchase, Long> {
